Split CI plan module
Move CI plan command/API, data structs, generated JSON support, and full job matrix definitions into a ci_plan directory while keeping the existing ci_plan.dart barrel import path.
