+++ tools/frb_internal/lib/src/makefile_dart/ci_plan/full_jobs.dart
@@ -1,0 +1,224 @@
+import 'structs.dart';
+
+const _githubHostedDesktopImages = [
+  'windows-2025',
+  'macos-15-intel',
+  'ubuntu-24.04',
+];
+
+const _exampleDartPackages = [
+  'frb_example--dart_minimal',
+  'frb_example--pure_dart',
+  'frb_example--pure_dart_pde',
+];
+
+const _flutterNativePackages = [
+  'frb_example--flutter_via_create',
+  'frb_example--flutter_package--example',
+  'frb_example--rust_ui_counter--ui',
+  'frb_example--rust_ui_todo_list--ui',
+];
+
+final kCiJobs = [
+  CiJob('deploy_website'),
+  CiJob('lint_rust_primary'),
+  CiJob('lint_dart_primary'),
+  CiJob('lint_rust_feature_flag'),
+  CiJob(
+    'generate_run_frb_codegen_command_generate',
+    matrix: CiMatrix([
+      for (final image in _githubHostedDesktopImages)
+        for (final package in [
+          'frb_example--dart_minimal',
+          'frb_example--pure_dart',
+          'frb_example--pure_dart_pde',
+          'frb_example--dart_build_rs',
+          'frb_example--deliberate_bad',
+          'frb_example--integrate_third_party',
+          'frb_example--flutter_via_create',
+          'frb_example--flutter_via_integrate',
+          'frb_example--flutter_package',
+          'frb_example--rust_ui_counter--ui',
+          'frb_example--rust_ui_todo_list--ui',
+        ])
+          if (!_isExcludedGenerateCommandGenerate(
+            image: image,
+            package: package,
+          ))
+            {'image': image, 'package': package},
+    ]),
+  ),
+  CiJob(
+    'generate_run_frb_codegen_command_integrate',
+    matrix: CiMatrix([
+      for (final image in ['macos-15-intel', 'windows-2025', 'ubuntu-24.04'])
+        for (final package in [
+          'frb_example--flutter_via_create',
+          'frb_example--flutter_via_integrate',
+          'frb_example--flutter_package',
+        ])
+          {'image': image, 'package': package, 'platforms': 'default'},
+      {
+        'image': 'ubuntu-24.04',
+        'package': 'frb_example--flutter_via_create',
+        'platforms': 'ohos',
+      },
+    ]),
+  ),
+  CiJob('generate_internal'),
+  CiJob(
+    'bench_dart_native',
+    matrix: CiMatrix([
+      for (final image in _githubHostedDesktopImages) {'image': image},
+    ]),
+  ),
+  CiJob('bench_upload'),
+  CiJob(
+    'build_flutter',
+    matrix: CiMatrix([
+      for (final info in [
+        {'image': 'windows-2025', 'target': 'windows'},
+        {'image': 'windows-11-arm', 'target': 'windows'},
+        {'image': 'macos-15-intel', 'target': 'macos'},
+        {'image': 'ubuntu-latest', 'target': 'linux'},
+        {'image': 'ubuntu-latest', 'target': 'android-aab'},
+        {'image': 'ubuntu-latest', 'target': 'android-apk'},
+        {'image': 'macos-15-intel', 'target': 'ios'},
+        {'image': 'ubuntu-latest', 'target': 'ohos'},
+      ])
+        {'info': info},
+    ]),
+  ),
+  CiJob(
+    'test_mimic_quickstart',
+    matrix: CiMatrix([
+      for (final image in ['windows-2025', 'macos-15-intel', 'ubuntu-latest'])
+        {'image': image},
+    ]),
+  ),
+  CiJob(
+    'test_rust',
+    matrix: CiMatrix([
+      for (final info in [
+        {'image': 'macos-15-intel', 'version': ''},
+        {'image': 'windows-2025', 'version': ''},
+        {'image': 'ubuntu-latest', 'version': ''},
+        {'image': 'ubuntu-latest', 'version': 'nightly'},
+        {'image': 'ubuntu-latest', 'version': '1.85.0'},
+      ])
+        {'info': info},
+    ]),
+  ),
+  CiJob(
+    'test_dart_native',
+    matrix: CiMatrix([
+      for (final image in _githubHostedDesktopImages)
+        for (final package in [
+          'frb_dart',
+          'frb_utils',
+          'tools--frb_internal',
+          ..._exampleDartPackages,
+          'frb_example--dart_build_rs',
+        ])
+          if (!_isExcludedTestDartNative(image: image, package: package))
+            {'image': image, 'package': package},
+    ]),
+  ),
+  CiJob(
+    'test_dart_web',
+    matrix: CiMatrix([
+      for (final package in ['frb_dart', ..._exampleDartPackages])
+        {'package': package},
+    ]),
+  ),
+  CiJob(
+    'test_dart_valgrind',
+    matrix: CiMatrix([
+      for (final package in _exampleDartPackages) {'package': package},
+    ]),
+  ),
+  CiJob(
+    'test_dart_sanitizer',
+    matrix: CiMatrix([
+      for (final sanitizer in ['asan', 'lsan'])
+        for (final package in _exampleDartPackages)
+          {'sanitizer': sanitizer, 'package': package},
+    ]),
+  ),
+  CiJob(
+    'test_flutter_native_android',
+    matrix: CiMatrix([
+      for (final package in _flutterNativePackages)
+        for (final device in ['pixel', 'Nexus 6'])
+          {'package': package, 'device': device, 'api-level': 35},
+    ]),
+  ),
+  CiJob(
+    'test_flutter_native_ios',
+    matrix: CiMatrix([
+      for (final package in _flutterNativePackages)
+        for (final device in [
+          'iPad (10th generation) Simulator (18.6)',
+          'iPhone 16 Pro Max Simulator (18.6)',
+        ])
+          {'package': package, 'device': device},
+    ]),
+  ),
+  CiJob(
+    'test_flutter_native_desktop',
+    matrix: CiMatrix([
+      for (final info in [
+        for (final package in _flutterNativePackages)
+          ..._flutterDesktopPackageEntries(package),
+        ..._linuxFlutterDesktopPackageEntries([
+          'frb_example--flutter_via_integrate',
+          'frb_example--gallery',
+          'frb_example--integrate_third_party',
+        ]),
+      ])
+        {'info': info},
+    ]),
+  ),
+  CiJob(
+    'test_flutter_web',
+    matrix: CiMatrix([
+      for (final package in [
+        'frb_example--flutter_via_create',
+        'frb_example--gallery',
+      ])
+        {'package': package},
+    ]),
+  ),
+  CiJob('misc_codecov'),
+];
+
+bool _isExcludedGenerateCommandGenerate({
+  required String image,
+  required String package,
+}) =>
+    (image == 'windows-2025' || image == 'macos-15-intel') &&
+    {
+      'frb_example--deliberate_bad',
+      'frb_example--integrate_third_party',
+      'frb_example--flutter_via_integrate',
+    }.contains(package);
+
+bool _isExcludedTestDartNative({
+  required String image,
+  required String package,
+}) =>
+    (image == 'windows-2025' || image == 'macos-15-intel') &&
+    {'frb_utils', 'tools--frb_internal'}.contains(package);
+
+List<Map<String, Object?>> _flutterDesktopPackageEntries(String package) => [
+  {'image': 'windows-2025', 'platform': 'windows', 'package': package},
+  {'image': 'macos-15-intel', 'platform': 'macos', 'package': package},
+  {'image': 'ubuntu-latest', 'platform': 'linux', 'package': package},
+];
+
+List<Map<String, Object?>> _linuxFlutterDesktopPackageEntries(
+  List<String> packages,
+) => [
+  for (final package in packages)
+    {'image': 'ubuntu-latest', 'platform': 'linux', 'package': package},
+];
